Create AWS Server Code
This script sends a file to an Amazon Web Services S3 Cloud Storage account where it will be stored. To know the upload speed of the servers, the code is sent rather than manually uploading. The access key and secret key is blocked at the moment for security purposes however, with the right credentials, a user can access this specific AWS account.